moved event management to its own component
